Update EBSVolumes.md
The replaced sentence was ambiguous, with the salient reading (for me, anyway) implying that a volume *can* attach to multiple instances if done with instances *outside* the Availability Zone. Removing the same-AZ condition from the sentence makes it read much more clearly, especially since the same-AZ condition is already stated in the previous paragraph. I've included a reiteration of that same-AZ condition in a separate, following, paragraph because I think it's important enough to deserve it, but you can judge whether it's needed. Thanks.