Update Upgrading_PV_drivers.md
I've added Windows Server 2022 to the list of windows server versions in the "Upgrade Windows Server instances (AWS PV upgrade)" section. This is assuming that the steps to upgrade AWS PV is the same for Windows Server 2022 as it is for the other listed server versions. I saw no indication elsewhere to indicate that there is a separate process for upgrading Windows Server 2022 and this article (https://docs.aws.amazon.com/AWSEC2/latest/WindowsGuide/xen-drivers-overview.html) mentions that Windows Server 2022 should be running latest AWS PV.